+++ graphs/quartet_test.go
@@ -1,4 +1,4 @@
-package qrt
+package graphs
 
 import (
 	"fmt"
